Add tests to Controllers

+++ server/src/main/java/ru/simple/chat/rest/ChatController.java
@@ -61,8 +61,11 @@ public class ChatController {
                                         @RequestParam(value = "name") String chatName,
                                         @RequestParam(value = "participants") List<User> participants) {
 
+        if (chatClient.contains(chatName)) {
+            return new ResponseEntity("Chat with name " + chatName + " already exists",
+                    HttpStatus.BAD_REQUEST);
+        }
         //TODO think about the faster variant this or userClient.contains() for every participant
-
         List<User> users = userClient.getAllUsers();
         Optional<User> badUser = participants.stream().filter(user -> !users.contains(user)).findFirst();
         if (badUser.isPresent()) {
@@ -86,14 +89,18 @@ public class ChatController {
      * @return the boolean
      */
     @DeleteMapping("/chat/{name}")
-    public ResponseEntity<?> deleteChat(@PathVariable("name") String chatName,
-                                        @RequestParam(value = "user") User author) {
+    public ResponseEntity deleteChat(@PathVariable("name") String chatName,
+                                     @RequestParam(value = "user") User author) {
         if (!chatClient.contains(chatName)) {
-            return new ResponseEntity<>("No chat with name " + chatName, HttpStatus.BAD_REQUEST);
+            return new ResponseEntity("No chat with name " + chatName, HttpStatus.BAD_REQUEST);
+        }
+
+        if (!chatClient.getChatByName(chatName).getAdmin().equals(author)) {
+            return new ResponseEntity("Only admin can delete chat", HttpStatus.BAD_REQUEST);
         }
 
         chatClient.deleteChat(chatName);
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     /**
